+++ src/web/assets/cp/src/js/BaseElementIndex.js
@@ -599,15 +599,13 @@ Craft.BaseElementIndex = Garnish.Base.extend({
             search: this.searchText,
             offset: this.settings.batchSize * (this.page - 1),
             limit: this.settings.batchSize,
+            drafts: this.drafts || null,
+            draftOf: false,
         };
 
-        // Only set trashed/drafts/draftOf params when needed, so we don't potentially override a source's criteria
+        // Only set trashed param when needed, so we don't potentially override a source's criteria
         if (this.trashed) {
             criteria.trashed = true;
-        }
-        if (this.drafts) {
-            criteria.drafts = true;
-            criteria.draftOf = false;
         }
 
         if (!Garnish.hasAttr(this.$source, 'data-override-status')) {